requirements.txt
dependency-injector==4.39.1
python-dotenv~=1.0.1
pydantic~=2.9.2
fastapi~=0.115.0
starlette~=0.38.6
